// apps/api/src/graphql/types.generated.d.ts
export type ApiResponse = {
    __typename?: 'ApiResponse';
    details?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
    message?: Maybe<Scalars['String']['output']>;
    ok?: Maybe<Scalars['Boolean']['output']>;
};
export type Assessment = {
    __typename?: 'Assessment';
    description?: Maybe<Scalars['String']['output']>;
    id: Scalars['ID']['output'];
    maxScore?: Maybe<Scalars['Float']['output']>;
    status?: Maybe<AssessmentStatus>;
    type?: Maybe<AssessmentType>;
    weight?: Maybe<Scalars['Float']['output']>;
};
export declare enum AssessmentStatus {
    Closed = "CLOSED",
    Draft = "DRAFT",
    Published = "PUBLISHED"
}
export declare enum AssessmentType {
    Assignment = "ASSIGNMENT",
    Exam = "EXAM",
    Oral = "ORAL",
    Practical = "PRACTICAL",
    Quiz = "QUIZ",
    Test = "TEST"
}
export type AttendanceRecord = {
    __typename?: 'AttendanceRecord';
    checkInTime?: Maybe<Scalars['DateTime']['output']>;
    date?: Maybe<Scalars['DateTime']['output']>;
    id?: Maybe<Scalars['String']['output']>;
    person?: Maybe<Person>;
    recordedBy?: Maybe<User>;
    status?: Maybe<AttendanceStatus>;
};
export type AttendanceSessionPayload = {
    __typename?: 'AttendanceSessionPayload';
    expiresAt: Scalars['DateTime']['output'];
    qrCodeDataUrl: Scalars['String']['output'];
    secretCode: Scalars['String']['output'];
    sessionId: Scalars['String']['output'];
};
export type AttendanceStats = {
    __typename?: 'AttendanceStats';
    absentCount: Scalars['Int']['output'];
    history?: Maybe<Array<DailyAttendance>>;
    lateCount: Scalars['Int']['output'];
    presentCount: Scalars['Int']['output'];
    rate: Scalars['Float']['output'];
    totalExpected: Scalars['Int']['output'];
};
export declare enum AttendanceStatus {
    Absent = "ABSENT",
    Excused = "EXCUSED",
    Late = "LATE",
    Present = "PRESENT"
}
export type Class = {
    __typename?: 'Class';
    _count?: Maybe<ClassCount>;
    defaultRoom?: Maybe<Room>;
    group?: Maybe<Group>;
    groupId: Scalars['ID']['output'];
    id: Scalars['ID']['output'];
    level: Scalars['String']['output'];
    name: Scalars['String']['output'];
    section?: Maybe<Scalars['String']['output']>;
    students?: Maybe<Array<Student>>;
    supervisor?: Maybe<Teacher>;
    teachingTeamMembers?: Maybe<Array<TeachingTeamMember>>;
    totalCoefficient?: Maybe<Scalars['Int']['output']>;
    totalWeeklyHours?: Maybe<Scalars['Float']['output']>;
};
export type ClassAndSubject = {
    __typename?: 'ClassAndSubject';
    classes?: Maybe<Array<Class>>;
    subjects?: Maybe<Array<Subject>>;
};
export type ClassCount = {
    __typename?: 'ClassCount';
    students?: Maybe<GenderStats>;
    subjects?: Maybe<Scalars['Int']['output']>;
    teachers?: Maybe<Scalars['Int']['output']>;
};
export type ClassList = {
    __typename?: 'ClassList';
    data?: Maybe<Array<Class>>;
    meta: PaginationMeta;
};
export type ClassStats = {
    __typename?: 'ClassStats';
    className: Scalars['String']['output'];
    studentCount: Scalars['Int']['output'];
};
// ...
export type ClassTeacher = {
    __typename?: 'ClassTeacher';
    classes?: Maybe<Array<Maybe<Class>>>;
    groups?: Maybe<Array<Group>>;
    schoolId?: Maybe<Scalars['ID']['output']>;
    teacher?: Maybe<Array<Maybe<Teacher>>>;
};
// ...
export type ResolversObject<TObject> = WithIndex<TObject>;
// ...
export type ApiResponseResolvers<ContextType = Context, ParentType extends ResolversParentTypes['ApiResponse'] = ResolversParentTypes['ApiResponse']> = ResolversObject<{
    details?: Resolver<Maybe<Array<Maybe<ResolversTypes['String']>>>, ParentType, ContextType>;
    message?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
    ok?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
}>;
export type AssessmentResolvers<ContextType = Context, ParentType extends ResolversParentTypes['Assessment'] = ResolversParentTypes['Assessment']> = ResolversObject<{
    description?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    maxScore?: Resolver<Maybe<ResolversTypes['Float']>, ParentType, ContextType>;
    status?: Resolver<Maybe<ResolversTypes['AssessmentStatus']>, ParentType, ContextType>;
    type?: Resolver<Maybe<ResolversTypes['AssessmentType']>, ParentType, ContextType>;
    weight?: Resolver<Maybe<ResolversTypes['Float']>, ParentType, ContextType>;
}>;
export type AttendanceRecordResolvers<ContextType = Context, ParentType extends ResolversParentTypes['AttendanceRecord'] = ResolversParentTypes['AttendanceRecord']> = ResolversObject<{
    checkInTime?: Resolver<Maybe<ResolversTypes['DateTime']>, ParentType, ContextType>;
    date?: Resolver<Maybe<ResolversTypes['DateTime']>, ParentType, ContextType>;
    id?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
    person?: Resolver<Maybe<ResolversTypes['Person']>, ParentType, ContextType>;
    recordedBy?: Resolver<Maybe<ResolversTypes['User']>, ParentType, ContextType>;
    status?: Resolver<Maybe<ResolversTypes['AttendanceStatus']>, ParentType, ContextType>;
}>;
export type AttendanceSessionPayloadResolvers<ContextType = Context, ParentType extends ResolversParentTypes['AttendanceSessionPayload'] = ResolversParentTypes['AttendanceSessionPayload']> = ResolversObject<{
    expiresAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
    qrCodeDataUrl?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    secretCode?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    sessionId?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
}>;
export type AttendanceStatsResolvers<ContextType = Context, ParentType extends ResolversParentTypes['AttendanceStats'] = ResolversParentTypes['AttendanceStats']> = ResolversObject<{
    absentCount?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
    history?: Resolver<Maybe<Array<ResolversTypes['DailyAttendance']>>, ParentType, ContextType>;
    lateCount?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
    presentCount?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
    rate?: Resolver<ResolversTypes['Float'], ParentType, ContextType>;
    totalExpected?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
}>;
export type ClassResolvers<ContextType = Context, ParentType extends ResolversParentTypes['Class'] = ResolversParentTypes['Class']> = ResolversObject<{
    _count?: Resolver<Maybe<ResolversTypes['ClassCount']>, ParentType, ContextType>;
    defaultRoom?: Resolver<Maybe<ResolversTypes['Room']>, ParentType, ContextType>;
    group?: Resolver<Maybe<ResolversTypes['Group']>, ParentType, ContextType>;
    groupId?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    level?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    section?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
    students?: Resolver<Maybe<Array<ResolversTypes['Student']>>, ParentType, ContextType>;
    supervisor?: Resolver<Maybe<ResolversTypes['Teacher']>, ParentType, ContextType>;
    teachingTeamMembers?: Resolver<Maybe<Array<ResolversTypes['TeachingTeamMember']>>, ParentType, ContextType>;
    totalCoefficient?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
    totalWeeklyHours?: Resolver<Maybe<ResolversTypes['Float']>, ParentType, ContextType>;
}>;
export type ClassAndSubjectResolvers<ContextType = Context, ParentType extends ResolversParentTypes['ClassAndSubject'] = ResolversParentTypes['ClassAndSubject']> = ResolversObject<{
    classes?: Resolver<Maybe<Array<ResolversTypes['Class']>>, ParentType, ContextType>;
    subjects?: Resolver<Maybe<Array<ResolversTypes['Subject']>>, ParentType, ContextType>;
}>;
export type ClassCountResolvers<ContextType = Context, ParentType extends ResolversParentTypes['ClassCount'] = ResolversParentTypes['ClassCount']> = ResolversObject<{
    students?: Resolver<Maybe<ResolversTypes['GenderStats']>, ParentType, ContextType>;
    subjects?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
    teachers?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
}>;
export type ClassListResolvers<ContextType = Context, ParentType extends ResolversParentTypes['ClassList'] = ResolversParentTypes['ClassList']> = ResolversObject<{
    data?: Resolver<Maybe<Array<ResolversTypes['Class']>>, ParentType, ContextType>;
    meta?: Resolver<ResolversTypes['PaginationMeta'], ParentType, ContextType>;
}>;
export type ClassStatsResolvers<ContextType = Context, ParentType extends ResolversParentTypes['ClassStats'] = ResolversParentTypes['ClassStats']> = ResolversObject<{
    className?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    studentCount?: Resolver<ResolversTypes['Int'], ParentType, ContextType>;
}>;
// ...
export type DayResolvers = EnumResolverSignature<{
    FRIDAY?: any;
    MONDAY?: any;
    SATURDAY?: any;
    SUNDAY?: any;
    THURSDAY?: any;
    TUESDAY?: any;
    WEDNESDAY?: any;
}, ResolversTypes['Day']>;
// ...
export type Resolvers<ContextType = Context> = ResolversObject<{
    ApiResponse?: ApiResponseResolvers<ContextType>;
    Assessment?: AssessmentResolvers<ContextType>;
    AttendanceRecord?: AttendanceRecordResolvers<ContextType>;
    AttendanceSessionPayload?: AttendanceSessionPayloadResolvers<ContextType>;
    AttendanceStats?: AttendanceStatsResolvers<ContextType>;
    Class?: ClassResolvers<ContextType>;
    ClassAndSubject?: ClassAndSubjectResolvers<ContextType>;
    ClassCount?: ClassCountResolvers<ContextType>;
    ClassList?: ClassListResolvers<ContextType>;
    ClassStats?: ClassStatsResolvers<ContextType>;
    ClassSubject?: ClassSubjectResolvers<ContextType>;
    ClassTeacher?: ClassTeacherResolvers<ContextType>;
    DailyAttendance?: DailyAttendanceResolvers<ContextType>;
    Date?: GraphQLScalarType;
    DateTime?: GraphQLScalarType;
    Day?: DayResolvers;
    Gender?: GenderResolvers;
    GenderStats?: GenderStatsResolvers<ContextType>;
    Group?: GroupResolvers<ContextType>;
    GroupType?: GroupTypeResolvers;
    Lesson?: LessonResolvers<ContextType>;
    LessonResources?: LessonResourcesResolvers<ContextType>;
    LessonStatus?: LessonStatusResolvers;
    LessonTeacher?: LessonTeacherResolvers<ContextType>;
    LessonsData?: LessonsDataResolvers<ContextType>;
    LessonsEvents?: LessonsEventsResolvers<ContextType>;
    LessonsList?: LessonsListResolvers<ContextType>;
    MonthlyRevenue?: MonthlyRevenueResolvers<ContextType>;
    MonthlyStats?: MonthlyStatsResolvers<ContextType>;
    Mutation?: MutationResolvers<ContextType>;
    PaginationMeta?: PaginationMetaResolvers<ContextType>;
    Parent?: ParentResolvers<ContextType>;
    ParentList?: ParentListResolvers<ContextType>;
    ParentStudent?: ParentStudentResolvers<ContextType>;
    Person?: PersonResolvers<ContextType>;
    Profile?: ProfileResolvers<ContextType>;
    Query?: QueryResolvers<ContextType>;
    RelationType?: RelationTypeResolvers;
    Room?: RoomResolvers<ContextType>;
    RoomList?: RoomListResolvers<ContextType>;
    School?: SchoolResolvers<ContextType>;
    SchoolId?: GraphQLScalarType;
    SchoolMembership?: SchoolMembershipResolvers<ContextType>;
    SchoolSettings?: SchoolSettingsResolvers<ContextType>;
    SchoolStats?: SchoolStatsResolvers<ContextType>;
    SearchClassesAndSubjects?: SearchClassesAndSubjectsResolvers<ContextType>;
    Staff?: StaffResolvers<ContextType>;
    Student?: StudentResolvers<ContextType>;
    StudentDisciplinaryAction?: StudentDisciplinaryActionResolvers<ContextType>;
    StudentList?: StudentListResolvers<ContextType>;
    StudentStatus?: StudentStatusResolvers;
    Subject?: SubjectResolvers<ContextType>;
    SubjectAssignments?: SubjectAssignmentsResolvers<ContextType>;
    SubjectCategory?: SubjectCategoryResolvers;
    SubjectList?: SubjectListResolvers<ContextType>;
    Teacher?: TeacherResolvers<ContextType>;
    TeacherAssignments?: TeacherAssignmentsResolvers<ContextType>;
    TeacherList?: TeacherListResolvers<ContextType>;
    TeachingTeamMember?: TeachingTeamMemberResolvers<ContextType>;
    TransportMode?: TransportModeResolvers;
    User?: UserResolvers<ContextType>;
    UserPayload?: UserPayloadResolvers<ContextType>;
}>;